src/helper.py
from ibapi.wrapper import EWrapper
from ibapi.client import EClient
from ibapi.connection import Connection
import logging
import time
import json

class MyTradingApp(EWrapper, EClient):
    def __init__(self):
        logging.info("Starting app")
        EClient.__init__(self, self)


    def connect_to_tws(self, host, port):
        try:
            logging.info("Connection to TWS(%s,%d)", host, port)
            self.connect(host, port, 0)
            time.sleep(5)
            if not self.isConnected():
                raise ConnectionError("helper: Failed to connect to TWS.")
        except Exception as e:
            logging.error(" Error connecting to TWS: %s", e)
            return False, str(e)

        # Add logic to handle connection (e.g., start a separate thread to run the message loop)
        return True, "Connected successfully"

    # Implement necessary methods from EWrapper
    # ...



